configuration: Pull out simple common types.
