+++ crates/configuration/src/lib.rs
@@ -1,5 +1,7 @@
 mod configuration;
 mod custom_trait_implementations;
+mod values;
+
 pub mod version1;
 pub mod version2;
 
@@ -7,6 +9,5 @@ pub use configuration::{
     as_runtime_configuration, configure, set_connection_uri, validate_raw_configuration,
     Configuration, RawConfiguration, RuntimeConfiguration,
 };
-pub use version2::{
-    occurring_scalar_types, ConnectionUri, IsolationLevel, PoolSettings, ResolvedSecret,
-};
+pub use values::{ConnectionUri, ResolvedSecret};
+pub use version2::{occurring_scalar_types, IsolationLevel, PoolSettings};
