+++ crates/configuration/src/values/mod.rs
@@ -1,0 +1,5 @@
+mod secret;
+mod uri;
+
+pub use secret::ResolvedSecret;
+pub use uri::ConnectionUri;
