+++ crates/tests/tests-common/src/common_tests/configuration_v1_tests.rs
@@ -3,6 +3,7 @@ use std::path::Path;
 
 use similar_asserts::assert_eq;
 
+use ndc_postgres_configuration as configuration;
 use ndc_postgres_configuration::version1;
 
 use crate::ndc_metadata::helpers::get_path_from_project_root;
@@ -23,8 +24,9 @@ pub async fn configure_is_idempotent(
     let mut args: version1::RawConfiguration = serde_json::from_value(expected_value.clone())
         .expect("Unable to deserialize as RawConfiguration");
 
-    args.connection_uri =
-        version1::ConnectionUri::Uri(version1::ResolvedSecret(connection_string.to_string()));
+    args.connection_uri = configuration::ConnectionUri::Uri(configuration::ResolvedSecret(
+        connection_string.to_string(),
+    ));
 
     let actual = version1::configure(args)
         .await
@@ -39,7 +41,7 @@ pub async fn configure_initial_configuration_is_unchanged(
     connection_string: &str,
 ) -> version1::RawConfiguration {
     let args = version1::RawConfiguration {
-        connection_uri: version1::ConnectionUri::Uri(version1::ResolvedSecret(
+        connection_uri: configuration::ConnectionUri::Uri(configuration::ResolvedSecret(
             connection_string.to_string(),
         )),
         ..version1::RawConfiguration::empty()
